refactor: split widget store presentation accessors
